docs: Add dependencies and installation instructions to README
This commit adds a new section to the README file with a list of dependencies required to run the script and instructions on how to install them using pip or conda. Additionally, a note has been added about the need to add Python to the PATH environment variable on Windows machines.

This change will make it easier for users to get set up and running the script without running into dependency-related issues.