Add dependencies and installation instructions for running the script
Added a section in the README file that lists the dependencies required to run the script and instructions on how to install them using pip or conda. Also included a note on adding Python to the PATH environment variable on Windows machines. This will help users get set up and running the script more easily.